# Conflicts:
#	core/src/main/java/io/github/infotest/character/Player.java
#	core/src/main/java/io/github/infotest/classes/Mage.java

+++ server_java/src/main/java/org/example/character/Player.java
@@ -49,7 +49,7 @@ public class Player {
     public Vector2 rotation;
 
     public String lastAttackedBy = "[Nothing]";
-    public int gold;
+    public float gold;
 
     public Player(String id, String name) {
         this.id = id;
